+++ base/strings/util.jl
@@ -359,9 +359,10 @@ _replace(io, repl, str, r, pattern) = print(io, repl)
 _replace(io, repl::Function, str, r, pattern) =
     print(io, repl(SubString(str, first(r), last(r))))
 
-function replace(str::String, pattern, repl, limit::Int)
-    limit == 0 && return str
-    limit < 0 && throw(DomainError())
+function replace_new(str::String, pattern, repl, count::Int)
+    # rename to `replace` when `replace` is removed from deprecated.jl
+    count == 0 && return str
+    count < 0 && throw(DomainError())
     n = 1
     e = endof(str)
     i = a = start(str)
@@ -386,7 +387,7 @@ function replace(str::String, pattern, repl, limit::Int)
         end
         r = search(str,pattern,k)
         j, k = first(r), last(r)
-        n == limit && break
+        n == count && break
         n += 1
     end
     write(out, SubString(str,i))
@@ -394,9 +395,9 @@ function replace(str::String, pattern, repl, limit::Int)
 end
 
 """
-    replace(string::AbstractString, pat, r, [count::Integer])
+    replace(s::AbstractString, pat, r, [count::Integer])
 
-Search for the given pattern `pat`, and replace each occurrence with `r`.
+Search for the given pattern `pat` in `s`, and replace each occurrence with `r`.
 If `count` is provided, replace at most `count` occurrences.
 As with search, the second argument may be a
 single character, a vector or a set of characters, a string, or a regular expression. If `r`
@@ -404,8 +405,7 @@ is a function, each occurrence is replaced with `r(s)` where `s` is the matched 
 If `pat` is a regular expression and `r` is a `SubstitutionString`, then capture group
 references in `r` are replaced with the corresponding matched text.
 """
-replace(s::AbstractString, pat, f, n::Integer=typemax(Int)) =
-    replace(String(s), pat, f, Int(n))
+replace(s::AbstractString, pat, f) = replace_new(String(s), pat, f, typemax(Int))
 
 # hex <-> bytes conversion